requirements.txt
pytest==4.0.1
coverage==4.5.2
pytest-flask==0.14.0
fam==2.0.5